[Moved] - Chrome api from Webext to utils folders in Extension Workspace
