+++ extension/utils/chrome-api.d.ts
